requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.5.1
azure-storage-blob==12.9.0
beautifulsoup4==4.10.0
urllib3==1.26.7
SQLAlchemy==1.4.25
pyodbc==4.0.32